Move minimongo modules
