+++ apps/meteor/client/cachedStores/PermissionsCachedStore.ts
@@ -1,6 +1,6 @@
 import type { IPermission } from '@rocket.chat/core-typings';
 
-import { PrivateCachedStore } from '../lib/cachedStores';
+import { PrivateCachedStore } from '../lib/cachedStores/CachedStore';
 import { Permissions } from '../stores';
 
 export const PermissionsCachedStore = new PrivateCachedStore<IPermission>({
