+++ apps/meteor/client/cachedStores/PrivateSettingsCachedStore.ts
@@ -1,7 +1,7 @@
 import type { ISetting } from '@rocket.chat/core-typings';
 
 import { sdk } from '../../app/utils/client/lib/SDKClient';
-import { PrivateCachedStore } from '../lib/cachedStores';
+import { PrivateCachedStore } from '../lib/cachedStores/CachedStore';
 import { PrivateSettings } from '../stores';
 
 class PrivateSettingsCachedStore extends PrivateCachedStore<ISetting> {
