--- apps/meteor/client/lib/cachedStores/index.ts
@@ -1,7 +1,0 @@
-export { CachedStoresManager } from './CachedStoresManager';
-export { pipe } from './pipe';
-export { applyQueryOptions } from './utils';
-export { createDocumentMapStore, type IDocumentMapStore } from './DocumentMapStore';
-export { MinimongoCollection } from './MinimongoCollection';
-export { PublicCachedStore, PrivateCachedStore } from './CachedStore';
-export { createGlobalStore } from './createGlobalStore';
