+++ apps/meteor/client/meteor/minimongo/ObserveHandle.ts
@@ -1,7 +1,7 @@
 import { Tracker } from 'meteor/tracker';
 
 import type { LocalCollection } from './LocalCollection';
-import type { Query } from './Query';
+import type { Query } from './queries';
 
 export class ObserveHandle<T extends { _id: string }> {
 	isReady: boolean;
