+++ apps/meteor/client/meteor/overrides/unstoreLoginToken.ts
@@ -1,6 +1,6 @@
 import { Accounts } from 'meteor/accounts-base';
 
-import { CachedStoresManager } from '../../lib/cachedStores';
+import { CachedStoresManager } from '../../lib/cachedStores/CachedStoresManager';
 
 const { _unstoreLoginToken } = Accounts;
 Accounts._unstoreLoginToken = (...args) => {
