+++ apps/meteor/client/providers/UserProvider/UserProvider.tsx
@@ -21,8 +21,8 @@ import { getUserPreference } from '../../../app/utils/client';
 import { sdk } from '../../../app/utils/client/lib/SDKClient';
 import { afterLogoutCleanUpCallback } from '../../../lib/callbacks/afterLogoutCleanUpCallback';
 import { useIdleConnection } from '../../hooks/useIdleConnection';
-import { applyQueryOptions } from '../../lib/cachedStores';
 import type { IDocumentMapStore } from '../../lib/cachedStores/DocumentMapStore';
+import { applyQueryOptions } from '../../lib/cachedStores/applyQueryOptions';
 import { createReactiveSubscriptionFactory } from '../../lib/createReactiveSubscriptionFactory';
 import { userIdStore } from '../../lib/userId';
 import { Users, Rooms, Subscriptions } from '../../stores';
