+++ apps/meteor/client/stores/Roles.ts
@@ -1,6 +1,7 @@
 import type { IRole } from '@rocket.chat/core-typings';
 
-import { createDocumentMapStore, createGlobalStore } from '../lib/cachedStores';
+import { createDocumentMapStore } from '../lib/cachedStores/DocumentMapStore';
+import { createGlobalStore } from '../lib/cachedStores/createGlobalStore';
 
 /** @deprecated prefer fetching data from the REST API, listening to changes via streamer events, and storing the state in a Tanstack Query */
 export const Roles = createGlobalStore(createDocumentMapStore<IRole>());
